pyproject.toml
[project]
name = "youtu-rag"
version = "1.0.0"
description = "Youtu-RAG is a next-generation agentic intelligent retrieval-augmented generation system based on the Autonomous Decision + Memory-Driven paradigm."
authors = [
    {name = "felixyfyu", email = "felixyfyu@tencent.com"}
]
license = "MIT"
keywords = ["agent", "ai", "llm", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "arize-phoenix>=12.19.0",
    "arize-phoenix-otel>=0.14.0",
    "art>=6.5",
    "arxiv>=2.2.0",
    "chromadb>=1.3.4",
    "chunkr-ai>=0.1.0",
    "colorlog>=6.9.0",
    "docker>=7.1.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.2",
    "fastmcp>=2.13.3",
    "google-genai>=1.26.0",
    "hydra-core>=1.3.2",
    "ipython>=9.5.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.3",
    "mcp>=1.12.3",
    # https://github.com/openai/openai-agents-python/issues/1395 | openai==1.97.1 & openai-agents==0.2.5
    "minio>=7.2.18",
    "openai-agents==0.3.3",
    # https://github.com/Arize-ai/openinference/tree/main/python/instrumentation/openinference-instrumentation-openai
    "openinference-instrumentation-openai>=0.1.30",
    "openinference-instrumentation-openai-agents>=1.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pexpect>=4.9.0",
    "pillow>=11.3.0",
    "prompt-toolkit>=3.0.52",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "pymupdf>=1.26.7",
    "pymysql>=1.1.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
    "requests>=2.32.4",
    "rich>=14.1.0",
    "sqlmodel>=0.0.24",
    "tiktoken>=0.9.0",
    "toml>=0.10.2",
    "typer>=0.17.4",
    "wikipedia-api>=0.8.1",
    "mysql-connector-python==9.5.0",
    "pymysql==1.1.2",
    "tabulate==0.9.0",
    "pip>=25.3",
    "numpy>=2.3.3",
    "langgraph>=1.0.5",
    "tqdm>=4.67.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "networkx>=3.5",
    "nltk>=3.9.1",
]

[project.urls]
Homepage = "https://github.com/TencentCloudADP/youtu-agent"
Documentation = "https://tencentcloudadp.github.io/youtu-agent/"
Repository = "https://github.com/TencentCloudADP/youtu-agent"

[project.optional-dependencies]
litllm = [
    "litellm>=1.77.0",
]

[dependency-groups]
dev = [
    "arize-phoenix-client>=1.14.1",
    "arize-phoenix-otel>=0.11.0",
    # datasets == 4.0.0 remove scripts, https://github.com/huggingface/datasets/issues/7693
    "datasets==3.6.0",
    "gradio>=5.42.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.16",
    "mkdocstrings[python]>=0.30.0",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.12.8",
    "tornado>=6.5.2",
    "twine>=6.2.0",
]
documents = [
    "pymupdf>=1.26.3",
]
gaia = [
    "astor>=0.8.1",
    "docx2markdown>=0.1.1",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "pymupdf>=1.26.3",
    "retry>=0.9.2",
    "tabulate>=0.9.0",
    "unstructured[docx,pptx]>=0.18.14",
    "validators>=0.35.0",
    "xls2xlsx>=0.2.0",
    "xmltodict>=0.15.0",
]
python-executor = [
    "chardet>=5.2.0",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
]
rl = [
    "aiolimiter>=1.2.1",
    "tencentcloud-sdk-python>=3.0.1474",
]
search = [
    "crawl4ai>=0.7.4",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["utu"]
exclude = ["utu/ui"]
# only-include = ["utu/utils"]


[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"

log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)s] %(name)s: %(message)s"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
addopts = [
    "-v", "-s",
]

[tool.uv]
index-url = "https://pypi.org/simple"
# extra-index-url = [
#     "https://mirrors.tencent.com/repository/pypi/tencent_pypi/simple",
# ]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    "**/*.pyi",
    ".*",
    "configs/",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
isort = { combine-as-imports = true, known-first-party = ["utu"] }

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.mypy]
strict = true
disallow_incomplete_defs = false
disallow_untyped_defs = false
disallow_untyped_calls = false

# [tool.pylint.messages_control]
# disable = ["broad-except", "bare-except", "invalid-name"]
